+++ src/components/common/pagination.tsx
@@ -1,6 +1,21 @@
 import Button from './button';
 import { PaginationType } from '@/src/type/orders/index';
-const Pagination = ({ currentPage, totalPages, nextPage, prevPage }: PaginationType) => {
+import { useEffect } from 'react';
+const Pagination = ({ data, pageSize, setCurrentPage, currentPage }: PaginationType) => {
+  const totalPages = Math.ceil(data.length / pageSize);
+
+  const nextPage = () => {
+    setCurrentPage(currentPage + 1);
+  };
+
+  const prevPage = () => {
+    setCurrentPage(currentPage - 1);
+  };
+
+  useEffect(() => {
+    window.scrollTo(0, 0);
+  }, [currentPage]);
+
   return (
     <div className='pagination'>
       <Button className='primary-button' buttonText='Previous' handleClick={prevPage} disable={currentPage === 1} />
